+++ src/main/java/com/atlauncher/data/minecraft/loaders/LoaderVersion.java
@@ -85,7 +85,7 @@ public class LoaderVersion {
 
     public boolean shouldInstallServerScripts() {
         // Forge 37 and newer (assumed) provide their own scripts for launching
-        if (this.isForge() && Integer.parseInt(this.version.substring(0, 2), 10) >= 37) {
+        if (this.isForge() && Integer.parseInt(this.version.substring(0, this.version.indexOf(".")), 10) >= 37) {
             return false;
         }