+++ CHANGELOG.md
@@ -16,5 +16,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - When selecting the bin folder of a Java install, correctly parse that for the launcher
 - Launcher not loading if cmd working directory doesn't exist [#518]
 - Logins not working for Microsoft accounts using game pass
+- Installer not working on Forge versions older than 10.x
 
 ### Misc